Remove tls_issuer_common
